Make asked changes
