+++ src/modules/APIs/comments.js
@@ -11,4 +11,17 @@ const getComments = async (pokemonID) => {
   }
 };
 
+export const postcommint = async (pokemonID, username, comment) => {
+  const requestOptions = {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({
+      item_id: pokemonID,
+      username,
+      comment,
+    }),
+  };
+  await fetch(`${involvementURL}?item_id=${pokemonID}`, requestOptions);
+};
+
 export default getComments;
